requirements.txt
asgiref==3.6.0
cloudinary==1.32.0
Django==3.2.18
django-cloudinary-storage==0.3.0
django-taggit==3.1.0
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
